+++ UI/_base/interface/IContext.ts
@@ -1,0 +1,6 @@
+export interface IContext {
+   scope: unknown;
+   get(field: string): Record<string, unknown>;
+   set(): void;
+   has(): boolean;
+}
